requirements.txt
# Smart City Streaming — Python Dependencies
# Install with: pip install -r requirements.txt

# Kafka client
confluent-kafka==2.3.0

# GPS / GPX parsing
gpxpy==1.6.2

# HTTP requests (OpenWeather API)
requests==2.31.0

# Env management
python-dotenv==1.0.1

# AWS SDK
boto3==1.34.69
botocore==1.34.69

# Spark (for local runs; cluster mode uses Docker image)
pyspark==3.5.1

# Fake data generation
Faker==24.4.0

# Utility
geopy==2.4.1    # Haversine distance helper
